src/apps/clipper/pyproject.toml
[tool.poetry]
name = "clipper"
version = "0.1.0"
description = ""
authors = ["voidfiles <voidfiles@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
httpx = "^0.27.0"
tenacity = "^8.4.2"
html-sanitizer = "^2.4.4"
readability-lxml = "^0.8.1"
markdownify = "^0.12.1"
jinja2 = "^3.1.4"
hishel = "^0.0.29"
markdown2 = "^2.4.13"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
flake8-black = "^0.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
readability = "readability:main"